Split up AlterOperation
